### 2026-03-15 #759 @thisisamirv
rfastlowess: High-Performance LOWESS Smoothing for R
Submitting Author Name: Amir Valizadeh
Submitting Author Github Handle: @thisisamirv
Other Package Authors Github handles:
Repository:  https://github.com/thisisamirv/lowess-project
Sub-directory: bindings/r
Version submitted: 1.2.0
Submission type: Standard

Version accepted: TBD

---

Package: rfastlowess
Title: High-Performance LOWESS Smoothing for R
Version: 1.2.0
Authors@R: person("Amir", "Valizadeh", email = "thisisamirv@gmail.com", role = c("aut", "cre", "fnd"), comment = c(ORCID = "0000-0001-5983-8527"))
Description: Provides high-performance LOWESS smoothing (Locally Weighted Scatterplot Smoothing)
    using a 'Rust' backend. Supports various weight functions, robustness iterations,
    streaming/online processing, cross-validation, and uncertainty quantification.
    Applicable to genomic data, time series, and general-purpose smoothing tasks.
License: MIT + file LICENSE | Apache License (== 2.0)
SystemRequirements: 'Rust' programming language (via 'rustup' or system tools)
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
biocViews: Preprocessing, StatisticalMethod, Regression, FunctionalPrediction, QualityControl, KEGG
Imports: BiocGenerics
Suggests: BiocStyle, knitr, rmarkdown, testthat (>= 3.0.0)
Config/rextendr/version: 0.4.2
```

    - [ ] rOpenSci internal tools

  This package wraps the high-performance Rust `lowess` and `fastLowess` crates to provide robust and extremely fast scatterplot smoothing. It serves as a data munging and transformation tool, allowing users to accurately extract trends and filter signals from noisy data series.

  The target audience includes statisticians, bioinformaticians, and data scientists. Scientific applications include genomic data smoothing, time series analysis, and general-purpose robust trend estimation and filtering.

  Yes, R provides `stats::lowess` and `stats::loess`. `rfastlowess` is significantly faster (on average 2-3x faster than R's base `lowess`), supports multithreaded parallel execution, boundary padding (to correct edge variance), MAD-based scale estimation for true breakdown-point robustness, auto-convergence, online streaming, and calculation of model confidence/prediction intervals, all of which standard implementations lack. It also supports different kernels, providing additional flexibility for users that require different smoothing properties.

  N/A. This package implements statistical algorithms and does not natively collect, manage, or require human subjects data.

  [Pre-submission inquiry](https://github.com/ropensci/software-review/issues/739)
  @emilyriederer

  N/A (all checks pass).

## Use of Generative AI
- [x] Generative AI tools were used to produce some of the material in this submission.

### Description of AI Usage

Generative AI (**Antigravity**, an agentic AI assistant) was used in different aspects of the production and maintenance of this software. Specifically:
- Design of the logo for the software.
- Writing the foundation of the bindings scripts.
- Identification and patching of security vulnerabilities (CVEs) in dependency stacks and hardening of FFI layers (Node.js, WASM).
- Maintaining complex Makefiles and ensuring cross-platform build consistency.
- Fixing broken workspace tests and improving test coverage.
- Research and help find different alternative solutions to performance aspects of the software.
- Design of different workflows for maintenance and release of the software.
- Debugging and organizing the codebase.

**Affirmation:** All AI-generated material has been carefully reviewed, tested, and approved by the human authors. Mathematical algorithms were specifically written by the author, cross-checked with the references, and validated for different scenarios.
