### 2026-03-18 #759 @thisisamirv
Thank you @emilyriederer . My apologies for forgetting to add that section, although it was mentioned in the pre-submission inquiry. I have edited my submission inquiry message and added that statement. A copy is provided below for your reference as well:

## Use of Generative AI
- [x] Generative AI tools were used to produce some of the material in this submission.

### Description of AI Usage

Generative AI (**Antigravity**, an agentic AI assistant) was used in different aspects of the production and maintenance of this software. Specifically:
- Design of the logo for the software.
- Writing the foundation of the bindings scripts.
- Identification and patching of security vulnerabilities (CVEs) in dependency stacks and hardening of FFI layers (Node.js, WASM).
- Maintaining complex Makefiles and ensuring cross-platform build consistency.
- Fixing broken workspace tests and improving test coverage.
- Research and help find different alternative solutions to performance aspects of the software.
- Design of different workflows for maintenance and release of the software.
- Debugging and organizing the codebase.

**Affirmation:** All AI-generated material has been carefully reviewed, tested, and approved by the human authors. Mathematical algorithms were specifically written by the author, cross-checked with the references, and validated for different scenarios.
