### 2026-05-03 #759 @maurolepore
The package is on our radar and we're still working to find a handling editor. I'll continue that work and follow up here.

Thanks for your patience and for sharing your work with the rOpenSci community.
