### 2026-05-03 #759 @maurolepore
@thisisamirv

I'm realizing that this issue used the standard template, but from the pre-submission conversation with @mpadge I understand this package is intended to be submitted for the stats review process. Can you both confirm what's the agreement?

If we move forward as stats, then [this comment](https://github.com/ropensci/software-review/issues/654#issuecomment-2350808047) suggests we should close this issue, do some homework, and open a new issue as stats

> ... after checking with the editorial board, I was alerted that you seemingly have mashed together our pre-submission and regular stats templates with your submission.
>
> Could I ask you to please close this and open a regular stats submission so that the bot picks up this as a stats submission?
>
> You will need to add our stats standard annotations to the package as well. The bot will pick that up and the process won't proceed until you've done that, so I suggest doing that before opening a new issue and using the proper stats template.
> -- @adamhsparks
