### 2026-05-05 #759 @thisisamirv
Hi @maurolepore, thank you for your responses. I do believe my package falls in the "stats" category, but I cannot say for sure, as I am not quite familiar with rOpenSci's internal policies and guidelines, and I also believe some policies changed while my pre-submission and main submission were ongoing. @mpadge, can you please confirm if I should resubmit this issue in a different category, or are we in the right place already?
